bi-chat/server/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
openai>=1.30.0
httpx>=0.27.0
pydantic>=2.5.0
sse-starlette>=1.8.0
mcp>=1.18.0
